Ensure we only call on deleted cooperation measures

+++ app/Console/Commands/Deploys/FixMeasureCommand.php
@@ -48,53 +48,57 @@ class FixMeasureCommand extends Command
         ];
 
         foreach ($deletedMeasures as $id => $name) {
-            $groupedAdvices = DB::table('user_action_plan_advices')
-                ->where('user_action_plan_advisable_type', CooperationMeasureApplication::class)
-                ->where('user_action_plan_advisable_id', $id)
-                ->get()
-                ->groupBy('user_id');
+            if (DB::table('cooperation_measure_applications')->where('id', $id)->doesntExist()) {
+                $groupedAdvices = DB::table('user_action_plan_advices')
+                    ->where('user_action_plan_advisable_type', CooperationMeasureApplication::class)
+                    ->where('user_action_plan_advisable_id', $id)
+                    ->get()
+                    ->groupBy('user_id');
 
-            foreach ($groupedAdvices as $userId => $advices) {
-                $building = DB::table('buildings')->where('user_id', $userId)->first();
+                foreach ($groupedAdvices as $userId => $advices) {
+                    $building = DB::table('buildings')->where('user_id', $userId)->first();
 
-                foreach ($advices as $advice) {
-                    // We don't need to worry about the measure not existing. While the advices didn't get updated
-                    // correctly, the measures were made for the right input sources.
-                    $customMeasure = DB::table('custom_measure_applications')
-                        ->where('name->nl', $name)
-                        ->where('building_id', $building->id)
-                        ->where('input_source_id', $advice->input_source_id)
-                        ->first();
+                    foreach ($advices as $advice) {
+                        // We don't need to worry about the measure not existing. While the advices didn't get updated
+                        // correctly, the measures were made for the right input sources.
+                        $customMeasure = DB::table('custom_measure_applications')
+                            ->where('name->nl', $name)
+                            ->where('building_id', $building->id)
+                            ->where('input_source_id', $advice->input_source_id)
+                            ->first();
 
-                    // Should always be the case, but just to be sure
-                    if ($customMeasure instanceof \stdClass) {
-                        $measureExists = DB::table('user_action_plan_advices')
-                            ->where('user_id', $userId)
-                            ->where('user_action_plan_advisable_type', CustomMeasureApplication::class)
-                            ->where('user_action_plan_advisable_id', $customMeasure->id)
-                            ->where('input_source_id', $customMeasure->input_source_id)
-                            ->exists();
+                        // Should always be the case, but just to be sure
+                        if ($customMeasure instanceof \stdClass) {
+                            $measureExists = DB::table('user_action_plan_advices')
+                                ->where('user_id', $userId)
+                                ->where('user_action_plan_advisable_type', CustomMeasureApplication::class)
+                                ->where('user_action_plan_advisable_id', $customMeasure->id)
+                                ->where('input_source_id', $customMeasure->input_source_id)
+                                ->exists();
 
-                        if ($measureExists) {
-                            Log::debug('Deleting advice with ID ' . $advice->id . ' from user ' . $userId . ' for deleted cooperation measure ' . $id);
+                            if ($measureExists) {
+                                Log::debug('Deleting advice with ID ' . $advice->id . ' from user ' . $userId . ' for deleted cooperation measure ' . $id);
 
-                            // Already copied over, so we can safely delete this one.
-                            DB::table('user_action_plan_advices')->where('id', $advice->id)->delete();
+                                // Already copied over, so we can safely delete this one.
+                                DB::table('user_action_plan_advices')->where('id', $advice->id)->delete();
+                            } else {
+                                Log::debug('Converting advice with ID ' . $advice->id . ' from user ' . $userId . ' for deleted cooperation measure ' . $id . ' to custom measure with ID ' . $customMeasure->id);
+
+                                // Not copied over yet. We must convert this advice to become the custom measure.
+                                DB::table('user_action_plan_advices')
+                                    ->where('id', $advice->id)
+                                    ->update([
+                                        'user_action_plan_advisable_type' => CustomMeasureApplication::class,
+                                        'user_action_plan_advisable_id' => $customMeasure->id,
+                                    ]);
+                            }
                         } else {
-                            Log::debug('Converting advice with ID ' . $advice->id . ' from user ' . $userId . ' for deleted cooperation measure ' . $id . ' to custom measure with ID ' . $customMeasure->id);
-
-                            // Not copied over yet. We must convert this advice to become the custom measure.
-                            DB::table('user_action_plan_advices')
-                                ->where('id', $advice->id)
-                                ->update([
-                                    'user_action_plan_advisable_type' => CustomMeasureApplication::class,
-                                    'user_action_plan_advisable_id' => $customMeasure->id,
-                                ]);
+                            Log::debug('No custom measure found for advice with ID ' . $advice->id . ' from user ' . $userId . ' for deleted cooperation measure ' . $id . ' and input source ' . $advice->input_source_id);
                         }
-                    } else {
-                        Log::debug('No custom measure found for advice with ID ' . $advice->id . ' from user ' . $userId . ' for deleted cooperation measure ' . $id . ' and input source ' . $advice->input_source_id);
                     }
                 }
+            } else {
+                Log::debug('Cooperation measure with ID ' . $id . ' exists! We\'re probably on accept.');
             }
         }
     }
